Add 3dsxtool build to haxe devkitarm.
There is a bug with 3dsxtool and larger binaries that is fixed in the
main tree but hasn't made a release yet because it is tied to the
devkitARM release cycle. This adds an updated 3dsxtool and causes haxe
compiled binaries to be able to be converted into .3dsx files.